Split HotspotInfo into lesser components, Make OverviewPage more structured
