+++ src/components/OverviewPage/Hotspot/HotspotContainer.css
