+++ src/components/OverviewPage/Hotspot/OpeningHours.css
