+++ src/components/OverviewPage/Hotspot/OpeningHoursTable.js
